requirements.txt
#Librerias Necesarias
pandas
pymongo
python-dotenv
jupyter